+++ webapp/src/App.css
@@ -36,3 +36,28 @@
     transform: rotate(360deg);
   }
 }
+
+input, button {
+  appearance: none;
+  background-color: transparent;
+  border: 0.13rem solid #0F0F0F;
+  border-radius: 15px;
+  box-sizing: border-box;
+  color: #0F0F0F;
+  padding: 1rem;
+  text-align: center;
+  text-decoration: none;
+  transition: all 300ms cubic-bezier(.23, 1, 0.32, 1);
+  user-select: none;
+  -webkit-user-select: none;
+  touch-action: manipulation;
+  width: 100%;
+  will-change: transform;
+}
+
+button:hover {
+  color: #fff;
+  background-color: #1A1A1A;
+  box-shadow: rgba(0, 0, 0, 0.25) 0 8px 15px;
+  transform: translateY(-2px);
+}
