+++ webapp/src/components/Login/Login.js
@@ -4,6 +4,7 @@ import axios from 'axios';
 import { Container, Typography, TextField, Button, Snackbar } from '@mui/material';
 import { useNavigate } from "react-router-dom";
 import useSignIn from 'react-auth-kit/hooks/useSignIn';
+import './Login.css';
 
 const Login = () => {
   const signIn = useSignIn();
@@ -51,47 +52,50 @@ const Login = () => {
   };
 
   return (
-    <Container component="main" maxWidth="xs" sx={{ marginTop: 4 }}>
+    <div className="login-container">
       {loginSuccess ? (
         <div>
-          <Typography component="h1" variant="h5" sx={{ textAlign: 'center' }}>
+          <h1 className="login-header">
             Hello {username}!
-          </Typography>
-          <Typography component="p" variant="body1" sx={{ textAlign: 'center', marginTop: 2 }}>
+          </h1>
+          <p className="login-text">
             Your account was created on {new Date(createdAt).toLocaleDateString()}.
-          </Typography>
+          </p>
         </div>
       ) : (
-        <div>
-          <Typography component="h1" variant="h5">
-            Login
-          </Typography>
-          <TextField
-            margin="normal"
-            fullWidth
-            label="Username"
+        <>
+          <h1 className="login-header">Login</h1>
+          <input
+            className="login-input"
+            type="text"
+            placeholder="Username"
             value={username}
             onChange={(e) => setUsername(e.target.value)}
           />
-          <TextField
-            margin="normal"
-            fullWidth
-            label="Password"
+          <input
+            className="login-input"
             type="password"
+            placeholder="Password"
             value={password}
             onChange={(e) => setPassword(e.target.value)}
           />
-          <Button variant="contained" color="primary" onClick={loginUser}>
+          <button className="login-button" onClick={loginUser}>
             Login
-          </Button>
-          <Snackbar open={openSnackbar} autoHideDuration={6000} onClose={handleCloseSnackbar} message="Login successful" />
-          {error && (
-            <Snackbar open={!!error} autoHideDuration={6000} onClose={() => setError('')} message={`Error: ${error}`} />
+          </button>
+          {openSnackbar && (
+            <div className="login-snackbar">
+              Login successful
+            </div>
           )}
-        </div>
+          {error && (
+            <div className="login-error">
+              Error: {error}
+            </div>
+          )}
+        </>
       )}
-    </Container>
-  );
-};
+      </div>
+    )
+}  
 
 export default Login;
