+++ webapp/src/pages/Authenticate/Authenticate.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
-import AddUser from '../components/AddUser';
-import Login from '../components/Login';
+import AddUser from '../../components/Register/Register';
+import Login from '../../components/Login/Login';
 import CssBaseline from '@mui/material/CssBaseline';
 import Container from '@mui/material/Container';
 import Typography from '@mui/material/Typography';
